Restructure of the backend

+++ crates/render/render_backend/render_backend_index/render_backend_index.h
@@ -2,4 +2,4 @@
 
 #include "./render_backend_index_core.h"
 
-typedef u32 RenderBackendIndex;
+typedef u16 RenderBackendIndex;
